fix: reqHeadrs.txt will contain headers to access APIs
